requirements.txt
adbutils>=2.8.0
opencv-python>=4.10.0
numpy>=1.26.0
pillow>=10.0.0
requests>=2.31.0
python-dotenv>=1.0.0
